+++ src/mock/mock.go
@@ -1,4 +1,4 @@
-package controller
+package mock
 
 import (
 	"net/http"
@@ -6,7 +6,6 @@ import (
 
 	"github.com/gin-gonic/gin"
 	"github.com/jphacks/A_2108/src/domain"
-	"github.com/jphacks/A_2108/src/mock"
 )
 
 func MockGetUserByID(c *gin.Context) {
@@ -22,9 +21,9 @@ func MockGetUserByID(c *gin.Context) {
 
 	var res domain.User
 	if planIdInt == 1 {
-		res = mock.MockUser1
+		res = MockUser1
 	} else if planIdInt == 2 {
-		res = mock.MockUser2
+		res = MockUser2
 	} else {
 		c.AbortWithStatusJSON(http.StatusBadRequest, gin.H{
 			"Error": "Not Found",
@@ -44,7 +43,7 @@ func MockGetPlanByID(c *gin.Context) {
 		return
 	}
 
-	for _, plan := range domain.MockPlans {
+	for _, plan := range MockPlans {
 		if plan.PlanId == planIdInt {
 			c.JSON(200, plan)
 			return
@@ -56,5 +55,5 @@ func MockGetPlanByID(c *gin.Context) {
 }
 
 func MockGetAllPlans(c *gin.Context) {
-	c.JSON(200, mock.MockPlans)
+	c.JSON(200, MockPlans)
 }
